pyproject.toml
[project]
name = "research-workspace"
version = "0.1.0"
description = "Research workspace for experiments"
requires-python = ">=3.10"
dependencies = [
    "openai>=2.16.0",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.53.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
